// src/app/shared/models/player.model.ts
export interface IPlayer {
  name: string;
  ponto1: number | null;
  ponto2: number | null;
  ponto3: number | null;
  ponto4: number | null;
  ponto5: number | null;
  ponto6: number | null;
  pontoS: number | null;
  pontoF: number | null;
  pontoP: number | null;
  pontoG: number | null;
  total: number;
  index: number;
}